requirements.txt
apify
playwright
python-dotenv
